Home vNext 1: restructure navigation (#2550)
* refactor: move Home as LegacyHome

* Temp navigation

* refactor: cleanup naming

* refactor: remove unused tabs setup

* header layout

* wip better container

* basic page structure respecting safe area

* Adjust back navigation

* refactor: navigation structure tweaks

* various tweaks

* self code review

* code review changes
