--- src/components/AndroidTabBarComponent/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "AndroidTabBarComponent.js",
-  "version": "0.0.0",
-  "private": true,
-  "main": "./AndroidTabBarComponent.js"
-}
