--- src/components/CustomTabBarComponent/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "CustomTabBarComponent.js",
-  "version": "0.0.0",
-  "private": true,
-  "main": "./CustomTabBarComponent.js"
-}
