+++ src/components/modern/Layout.js
@@ -1,0 +1,55 @@
+// @flow
+/*
+    Pillar Wallet: the personal data locker
+    Copyright (C) 2021 Stiftung Pillar Project
+
+    This program is free software; you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation; either version 2 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License along
+    with this program; if not, write to the Free Software Foundation, Inc.,
+    51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+*/
+
+import * as React from 'react';
+import { ScrollView } from 'react-native';
+import SafeAreaView from 'react-native-safe-area-view';
+import styled from 'styled-components/native';
+
+// Types
+import type { ViewProps } from 'utils/types/react-native';
+
+/**
+ * Root element for screens. Normally contains `HeaderBlock` & `Content`.
+ */
+export const Container: React.ComponentType<ViewProps> = styled.View`
+  flex: 1;
+`;
+
+type ContentProps = {|
+  children: React.Node,
+|};
+
+/**
+ * Content for regular screens, holds in children inside `ScrollView`.
+ *
+ * Correctly handles safe area.
+ */
+export function Content({ children }: ContentProps) {
+  return (
+    <ScrollView contentContainerStyle={{ flexGrow: 1 }}>
+      <ContentSafeArea>{children}</ContentSafeArea>
+    </ScrollView>
+  );
+}
+
+const ContentSafeArea = styled(SafeAreaView)`
+  flex: 1;
+`;
