+++ src/screens/LegacyHome/RariPoolItem.js
