+++ src/screens/LegacyHome/WalletsPart.js
