+++ src/screens/LegacyHome/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "LegacyHome",
+  "version": "0.0.0",
+  "private": true,
+  "main": "./LegacyHome.js"
+}
